refactor : create authentication class to manage the finger print authentication
