+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
             android:value="@string/google_maps_key"/>
         <activity
 
-            android:name=".presentation.MainActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.SPAplication.NoActionBar">
